Move notification handling to notification manager
