+++ EventPlanner/app/src/main/java/m3/eventplanner/utils/NotificationWebSocketManager.java
@@ -1,14 +1,26 @@
 package m3.eventplanner.utils;
 
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
 import android.content.Context;
+import android.os.Build;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
+import androidx.core.app.NotificationCompat;
 
 import com.google.gson.Gson;
 
 import java.util.function.Consumer;
 
 import m3.eventplanner.BuildConfig;
+import m3.eventplanner.R;
+import m3.eventplanner.auth.TokenManager;
+import m3.eventplanner.clients.ClientUtils;
 import m3.eventplanner.models.GetNotificationDTO;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 import ua.naiksoftware.stomp.Stomp;
 import ua.naiksoftware.stomp.StompClient;
 import ua.naiksoftware.stomp.dto.LifecycleEvent;
@@ -53,6 +65,48 @@ public class NotificationWebSocketManager {
 
             stompClient.connect();
         }
+    }
+
+    public static void showNotification(Context context, GetNotificationDTO notification) {
+        Integer accountId = new TokenManager(context).getAccountId();
+        ClientUtils clientUtils = new ClientUtils(context);
+
+        clientUtils.getNotificationService().isNotificationSilenced(accountId).enqueue(new Callback<Boolean>() {
+            @Override
+            public void onResponse(@NonNull Call<Boolean> call, @NonNull Response<Boolean> response) {
+                if (response.isSuccessful() && Boolean.FALSE.equals(response.body())) {
+                    displayLocalNotification(context, notification);
+                }
+            }
+
+            @Override
+            public void onFailure(@NonNull Call<Boolean> call, @NonNull Throwable t) {
+                t.printStackTrace();
+            }
+        });
+    }
+
+    private static void displayLocalNotification(Context context, GetNotificationDTO notification) {
+        NotificationManager notificationManager =
+                (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+
+        String channelId = "notifications_channel";
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            NotificationChannel channel = new NotificationChannel(
+                    channelId,
+                    "Notifications",
+                    NotificationManager.IMPORTANCE_DEFAULT
+            );
+            notificationManager.createNotificationChannel(channel);
+        }
+
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(context, channelId)
+                .setSmallIcon(R.drawable.ic_bell)
+                .setContentTitle(notification.getTitle())
+                .setContentText(notification.getContent())
+                .setAutoCancel(true);
+
+        notificationManager.notify(notification.getId(), builder.build());
     }
